optimize description
